Document missing useLayoutEffect popper dependency
DropdownMenu uses `useLayoutEffect` to update `popper`. If you want to make ESLint happy and add `popper` as a `useLayoutEffect` hook dependency this causes an endless loop because `popper` object changes on every render.

It makes sense to have this in the documentation because it is painful to debug why the app freezes when you toggle the menu.